+++ products/DesignerApp.browser.js
@@ -169,7 +169,7 @@ class DesignerApp {
     const errs = this.grammarProgram.getAllErrors().map(err => err.toObject())
     this._grammarErrorsConsole.html(errs.length ? new jtree.TreeNode(errs).toFormattedTable(200) : "0 errors")
     const grammarProgram = new jtree.GrammarProgram(this.grammarInstance.getValue())
-    this._readmeComponent.html(grammarProgram.toReadMe())
+    this._readmeComponent.html(new dumbdownNode(grammarProgram.toReadMe()).compile())
   }
   _updateShareLink() {
     const url = new URL(location.href)